Refactor experimental display preferences structure
