+++ src/apps/experimental/features/preferences/components/ItemDetailPreferences.tsx
@@ -7,7 +7,8 @@ import Typography from '@mui/material/Typography';
 import React from 'react';
 
 import globalize from 'lib/globalize';
-import { DisplaySettingsValues } from './types';
+
+import type { DisplaySettingsValues } from '../types/displaySettingsValues';
 
 interface ItemDetailPreferencesProps {
     onChange: (event: React.SyntheticEvent) => void;
