+++ src/apps/experimental/features/preferences/constants/locales.ts
